Initial code commit
This commit is a merged version of all my previous failed Mattrax projects. This restart is designed to take into account all of what I have learnt through the failures of my other Mattrax projects to finally build a working, deployable and useful version of Mattrax. This code uses sqlc for its datastore to PostgreSQL. Windows MDM enrollment is working but the AzureAD authentication provider is the only one currently functioning.