feat: Added device inventory
Added the device inventory which holds information about the state of the device. This data can then be parsed for the future web API. When enrolling a device the new 'deploying' state is used to detect failures in the final step of the enrollment process. The state is changed to 'managed' on the first sync.