refactor: Specific data format handlers exposed
Parsing and generation of SOAP, SyncML and WAP Provisioning Profile moved into the public code directory so it can be imported by other Go projects if needed. This move also removed Zerolog from these packages in preference of a function which can be used to support any logging or error tracing framework.